LPS-156422 Add index file to SupportOverview component
