+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/SupportOverview/SupportOverview.js
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import SupportOverviewSkeleton from './Skeleton';
+import LiferayContact from './components/LiferayContact';
+import SlaCardList from './components/SlaCardList';
+
+const SupportOverview = ({koroneikiAccount}) => (
+	<div className="container cp-project-contacs-container mb-5 mx-0">
+		<div className="row">
+			<div className="col-5 pb-8">
+				<SlaCardList koroneikiAccount={koroneikiAccount} />
+			</div>
+
+			<div className="col-7">
+				<LiferayContact koroneikiAccount={koroneikiAccount} />
+			</div>
+		</div>
+	</div>
+);
+
+SupportOverview.Skeleton = SupportOverviewSkeleton;
+
+export default SupportOverview;
